pyproject.toml
[project]
name = "kbm"
version = "0.1.0"
description = "Persistent memory for LLMs, accessible via MCP."
readme = "README.md"

# MARK: - Dependencies

requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0b1",
    "typer>=0.21",
    "pydantic-settings>=2.12",
    "raganything[all]>=1.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1",
    "pytest-asyncio>=1.3",
    "ruff>=0.14",
]

# MARK: - Build System

[project.scripts]
kbm = "kbm:app"

[build-system]
requires = ["uv_build>=0.9,<0.10"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["kbm"]

# MARK: - Tool Configuration

[tool.ruff]
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "I", "W", "RUF"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
useLibraryCodeForTypes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
